+++ crates/types-traits/interfaces/src/connector_types.rs
@@ -106,6 +106,13 @@ pub trait ConnectorServiceTrait<T: PaymentMethodDataTypes>:
     + MandateRevokeV2
     + VerifyWebhookSourceV2
     + VerifyRedirectResponse
+{
+}
+
+pub trait SurchargeServiceTrait: ConnectorCommon + SurchargeCalculateV2 {}
+
+pub trait PayoutServiceTrait:
+    ConnectorCommon
     + PayoutCreateV2
     + PayoutTransferV2
     + PayoutGetV2
@@ -116,8 +123,6 @@ pub trait ConnectorServiceTrait<T: PaymentMethodDataTypes>:
     + PayoutEnrollDisburseAccountV2
 {
 }
-
-pub trait SurchargeServiceTrait: ConnectorCommon + SurchargeCalculateV2 {}
 
 pub trait PaymentVoidV2:
     ConnectorIntegrationV2<connector_flow::Void, PaymentFlowData, PaymentVoidData, PaymentsResponseData>
@@ -137,6 +142,8 @@ pub trait PaymentVoidPostCaptureV2:
 pub type BoxedConnector<T> = Box<&'static (dyn ConnectorServiceTrait<T> + Sync)>;
 
 pub type BoxedSurchargeConnector = Box<&'static (dyn SurchargeServiceTrait + Sync)>;
+
+pub type BoxedPayoutConnector = Box<&'static (dyn PayoutServiceTrait + Sync)>;
 
 pub trait ValidationTrait: ConnectorCommon {
     fn should_do_order_create(&self) -> bool {